TractSeg_presentation_maker/requirements.txt
python-pptx
Pillow
vtk